vastly simplify component logic
